Mount command added for setup of component tests with C8yFetchClient. oauthLogin now in separate file. Refactoring of dependencies.
