+++ package.json
@@ -39,15 +39,12 @@
     "set-cookie-parser": "^2.6.0"
   },
   "devDependencies": {
-    "@angular/common": "^14.1.2",
-    "@c8y/client": "^1017.0.374",
     "@types/chai": "^4.3.4",
     "@types/glob": "^8.1.0",
     "@types/lodash": "^4.14.202",
     "@types/node": "^20.10.6",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
-    "cypress": "^13.3.2",
     "http-server": "^14.1.1",
     "rimraf": "^5.0.1",
     "start-server-and-test": "^2.0.0",
@@ -55,7 +52,8 @@
     "yalc": "^1.0.0-pre.53"
   },
   "peerDependencies": {
-    "@angular/common": ">=14.0.0",
+    "@angular/common": ">=16.0.0",
+    "@angular/core": ">=16.0.0",
     "@c8y/client": ">=1015.0.0",
     "cypress": ">=12.0.0",
     "lodash": ">=4.0.0"
