+++ packages/pactrunner/package.json
@@ -39,11 +39,6 @@
     "@types/chai": "^4.3.4",
     "@types/glob": "^8.1.0",
     "@types/node": "^20.10.6",
-    "copyfiles": "^2.4.1",
-    "cypress": "^13.3.2",
-    "http-server": "^14.1.1",
-    "rimraf": "^5.0.1",
-    "start-server-and-test": "^2.0.0",
     "typescript": "^5.0.2"
   }
 }
