+++ src/lib/commands/login.ts
@@ -1,8 +1,8 @@
+import { C8yAuthOptions } from "./auth";
+
 const { _ } = Cypress;
 const semver = require("semver");
 const { getAuthOptions, isAuth, resetClient } = require("./../utils");
-var setCookieParser = require("set-cookie-parser");
-const { getBaseUrlFromEnv } = require("../utils");
 
 declare global {
   namespace Cypress {
@@ -41,11 +41,6 @@ declare global {
         auth: C8yAuthOptions,
         options?: C8yLoginOptions
       ): Chainable<C8yAuthOptions>;
-
-      oauthLogin(): Chainable<C8yAuthOptions>;
-      oauthLogin(auth: C8yAuthOptions): Chainable<C8yAuthOptions>;
-      oauthLogin(user: string): Chainable<C8yAuthOptions>;
-      oauthLogin(user: string, password: string): Chainable<C8yAuthOptions>;
     }
   }
 
@@ -157,75 +152,5 @@ Cypress.Commands.add(
       }
     );
     cy.wrap(auth, { log: false });
-  }
-);
-
-Cypress.Commands.add(
-  "oauthLogin",
-  // @ts-ignore
-  { prevSubject: "optional" },
-  function (...args: C8yLoginAuthArgs) {
-    const auth: C8yAuthOptions = getAuthOptions(...args);
-    expect(auth).to.not.be.undefined;
-
-    const consoleProps: any = {};
-    Cypress.log({
-      name: "oauthLogin",
-      message: auth,
-      consoleProps: () => consoleProps,
-    });
-    if (Cypress.testingType === "component" && !getBaseUrlFromEnv()) {
-      const error = new Error(
-        "Base URL not set. Use C8Y_BASEURL env variable for component testing."
-      );
-      error.name = "C8yPactError";
-      throw error;
-    }
-    return (Cypress.isCy(auth) ? auth : cy.wrap(auth, { log: false })).then(
-      (auth: C8yAuthOptions) => {
-        consoleProps.auth = auth;
-
-        const tenant: string = auth.tenant || Cypress.env("C8Y_TENANT");
-        consoleProps.tenant = tenant;
-        cy.request({
-          method: "POST",
-          url: `${getBaseUrlFromEnv() || ""}/tenant/oauth?tenant_id=${tenant}`,
-          body: {
-            grant_type: "PASSWORD",
-            username: auth.user,
-            password: auth.password,
-            tfa_code: undefined,
-          },
-          form: true,
-        }).then((response) => {
-          let setCookie = response.headers.getSetCookie;
-          let cookieHeader: string[];
-          if (_.isFunction(response.headers.getSetCookie)) {
-            cookieHeader = (response.headers.getSetCookie as () => string[])();
-          } else if (_.isString(setCookie)) {
-            cookieHeader = [setCookie];
-          } else if (_.isArray(setCookie)) {
-            cookieHeader = setCookie;
-          } else {
-            cookieHeader = response.headers["set-cookie"] as string[];
-          }
-
-          setCookieParser(cookieHeader || []).forEach((c: any) => {
-            cy.setCookie(c.name, c.value);
-            if (_.isEqual(c.name.toLowerCase(), "authorization")) {
-              auth.bearer = c.value;
-            }
-            if (_.isEqual(c.name.toLowerCase(), "xsrf-token")) {
-              auth.xsrfToken = c.value;
-            }
-          });
-
-          Cypress.env("C8Y_LOGGED_IN_USER", auth.user);
-          Cypress.env("C8Y_LOGGED_IN_USER_ALIAS", auth.userAlias);
-
-          return cy.wrap(auth, { log: false });
-        });
-      }
-    );
   }
 );
