+++ src/lib/pact/fetchclient.ts
@@ -11,6 +11,7 @@ import {
   wrapFetchResponse,
 } from "../../shared/c8yclient";
 import { C8yDefaultPact } from "../../shared/c8ypact";
+import { C8yAuthOptions } from "../commands/auth";
 const { getAuthOptions, getBaseUrlFromEnv } = require("./../utils");
 
 const { _ } = Cypress;
